+++ Tests/MapboxMapsTests/Style/StyleColorTests.swift
@@ -8,17 +8,17 @@ final class StyleColorTests: XCTestCase {
     var blue: Double!
     var alpha: Double!
     var colorSpaces: [CGColorSpace]!
+    var hue: Double!
+    var saturation: Double!
+    var lightness: Double!
 
-    var rgbaString: String {
-        "rgba(\(red!), \(green!), \(blue!), \(alpha!))"
-    }
+    var rgbString: String { String(format: "rgb(%.2f, %.2f, %.2f)", red, green, blue) }
+    var rgbaString: String { String(format: "rgba(%.2f, %.2f, %.2f, %.2f)", red, green, blue, alpha) }
+    var hslString: String { String(format: "hsl(%.2f, %.2f, %.2f)", hue, saturation, lightness) }
+    var hslaString: String { String(format: "hsla(%.2f, %.2f, %.2f, %.2f)", hue, saturation, lightness, alpha) }
 
     override func setUp() {
         super.setUp()
-        red = .random(in: 0...255)
-        green = .random(in: 0...255)
-        blue = .random(in: 0...255)
-        alpha = .random(in: 0...1)
         colorSpaces = [
             CGColorSpace.genericCMYK,
             CGColorSpace.displayP3,
@@ -40,6 +40,13 @@ final class StyleColorTests: XCTestCase {
             CGColorSpace.linearGray,
             CGColorSpace.extendedLinearGray,
         ].map { CGColorSpace(name: $0)! }
+        red = 83
+        green = 67
+        blue = 219
+        hue = 93
+        saturation = 37
+        lightness = 8
+        alpha = 0.5
     }
 
     override func tearDown() {
@@ -48,59 +55,66 @@ final class StyleColorTests: XCTestCase {
         green = nil
         red = nil
         colorSpaces = nil
+        hue = nil
+        saturation = nil
+        lightness = nil
         super.tearDown()
     }
 
-    func verify(_ styleColor: StyleColor?, line: UInt = #line) {
-        XCTAssertEqual(styleColor?.red, red, line: line)
-        XCTAssertEqual(styleColor?.green, green, line: line)
-        XCTAssertEqual(styleColor?.blue, blue, line: line)
-        XCTAssertEqual(styleColor?.alpha, alpha, line: line)
-    }
-
-    func invalidComponentValue(outsideOf range: ClosedRange<Double>) -> Double {
-        precondition(range.lowerBound != -.greatestFiniteMagnitude)
-        precondition(range.upperBound != .greatestFiniteMagnitude)
-        return Bool.random()
-            ? .random(in: -.greatestFiniteMagnitude...range.lowerBound)
-            : .random(in: range.upperBound...(.greatestFiniteMagnitude))
-    }
-
     func testComponentWiseInit() {
-        verify(StyleColor(red: red, green: green, blue: blue, alpha: alpha))
+        XCTAssertEqual(StyleColor(red: red, green: green, blue: blue)?.rawValue, rgbString)
+        XCTAssertEqual(StyleColor(red: red, green: green, blue: blue, alpha: alpha)?.rawValue, rgbaString)
+        XCTAssertEqual(StyleColor(hue: hue, saturation: saturation, lightness: lightness)?.rawValue, hslString)
+        XCTAssertEqual(StyleColor(hue: hue, saturation: saturation, lightness: lightness, alpha: alpha)?.rawValue, hslaString)
     }
 
     func testComponentWiseInitFailure() {
-        XCTAssertNil(StyleColor(red: invalidComponentValue(outsideOf: 0...255), green: 0, blue: 0, alpha: 0))
-        XCTAssertNil(StyleColor(red: 0, green: invalidComponentValue(outsideOf: 0...255), blue: 0, alpha: 0))
-        XCTAssertNil(StyleColor(red: 0, green: 0, blue: invalidComponentValue(outsideOf: 0...255), alpha: 0))
-        XCTAssertNil(StyleColor(red: 0, green: 0, blue: 0, alpha: invalidComponentValue(outsideOf: 0...1)))
+        for invalidComponent in [-1.0, 10000.0] {
+            XCTAssertNil(StyleColor(red: invalidComponent, green: 0, blue: 0))
+            XCTAssertNil(StyleColor(red: 0, green: invalidComponent, blue: 0))
+            XCTAssertNil(StyleColor(red: 0, green: 0, blue: invalidComponent))
+
+            XCTAssertNil(StyleColor(red: invalidComponent, green: 0, blue: 0, alpha: 0))
+            XCTAssertNil(StyleColor(red: 0, green: invalidComponent, blue: 0, alpha: 0))
+            XCTAssertNil(StyleColor(red: 0, green: 0, blue: invalidComponent, alpha: 0))
+            XCTAssertNil(StyleColor(red: 0, green: 0, blue: 0, alpha: invalidComponent))
+
+            XCTAssertNil(StyleColor(hue: invalidComponent, saturation: 0, lightness: 0))
+            XCTAssertNil(StyleColor(hue: 0, saturation: invalidComponent, lightness: 0))
+            XCTAssertNil(StyleColor(hue: 0, saturation: 0, lightness: invalidComponent))
+
+            XCTAssertNil(StyleColor(hue: invalidComponent, saturation: 0, lightness: 0, alpha: 0))
+            XCTAssertNil(StyleColor(hue: 0, saturation: invalidComponent, lightness: 0, alpha: 0))
+            XCTAssertNil(StyleColor(hue: 0, saturation: 0, lightness: invalidComponent, alpha: 0))
+            XCTAssertNil(StyleColor(hue: 0, saturation: 0, lightness: 0, alpha: invalidComponent))
+        }
     }
 
     func testUIColorInit() {
-        // use local input values since the ones from setUp are in the wrong range for this test
-        let red = CGFloat.random(in: 0...1)
-        let green = CGFloat.random(in: 0...1)
-        let blue = CGFloat.random(in: 0...1)
-        let alpha = CGFloat.random(in: 0...1)
+        let rgbaColor = StyleColor(UIColor(red: red / 255, green: green / 255, blue: blue / 255, alpha: alpha))
+        let hslaColor = StyleColor(UIColor(hue: hue / 360.0, saturation: saturation / 100.0, brightness: lightness / 100.0, alpha: alpha))
 
-        let color = StyleColor(UIColor(red: red, green: green, blue: blue, alpha: alpha))
-
-        XCTAssertEqual(color.red, Double(red * 255), accuracy: 0.001)
-        XCTAssertEqual(color.green, Double(green * 255), accuracy: 0.001)
-        XCTAssertEqual(color.blue, Double(blue * 255), accuracy: 0.001)
-        XCTAssertEqual(color.alpha, Double(alpha), accuracy: 0.001)
+        XCTAssertEqual(rgbaColor.rawValue, rgbaString)
+        XCTAssertEqual(hslaColor.rawValue, "rgba(16.25, 20.40, 12.85, 0.50)")
     }
 
     func testExpressionInit() {
-        let expression = Exp(.rgba) {
-            red!
-            green!
-            blue!
-            alpha!
-        }
+        let expressions = [
+            Exp(.rgb) { red!; green!; blue! },
+            Exp(.rgba) { red!; green!; blue!; alpha! },
+            Exp(.hsl) { hue!; saturation!; lightness! },
+            Exp(.hsla) { hue!; saturation!; lightness!; alpha! }
+        ]
+        let expectedColorStrings = [
+            rgbString,
+            rgbaString,
+            hslString,
+            hslaString
+        ]
 
-        verify(StyleColor(expression: expression))
+        for (expression, colorString) in zip(expressions, expectedColorStrings) {
+            XCTAssertEqual(StyleColor(expression: expression)?.rawValue, colorString)
+        }
     }
 
     func testExpressionInitFailureWrongOperator() {
@@ -114,124 +128,148 @@ final class StyleColorTests: XCTestCase {
     }
 
     func testExpressionInitFailureValuesOutOfBounds() {
-        func makeExpression(red: Double = 0, green: Double = 0, blue: Double = 0, alpha: Double = 0) -> Expression {
-            Exp(.rgba) {
-                red
-                green
-                blue
-                alpha
+        let invalidComponents = [
+            [0, 0, -1],
+            [0, 0, 1000]
+        ].map { $0.map(Exp.Argument.number) }
+
+        for var components in invalidComponents {
+            for _ in 0..<components.count {
+                XCTAssertNil(StyleColor(expression: Exp(operator: .rgb, arguments: components)))
+                XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: components + [.number(0)])))
+                XCTAssertNil(StyleColor(expression: Exp(operator: .hsl, arguments: components)))
+                XCTAssertNil(StyleColor(expression: Exp(operator: .hsla, arguments: components + [.number(0)])))
+
+                // shift the invalid component by one to the left to test all of them
+                components = components.rotatingLeft(positions: 1)
             }
         }
-        XCTAssertNil(StyleColor(expression: makeExpression(red: invalidComponentValue(outsideOf: 0...255))))
-        XCTAssertNil(StyleColor(expression: makeExpression(green: invalidComponentValue(outsideOf: 0...255))))
-        XCTAssertNil(StyleColor(expression: makeExpression(blue: invalidComponentValue(outsideOf: 0...255))))
-        XCTAssertNil(StyleColor(expression: makeExpression(alpha: invalidComponentValue(outsideOf: 0...1))))
+
+        // tests handling of invalid alpha component
+        XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: [.number(0), .number(0), .number(0), .number(-1)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: [.number(0), .number(0), .number(0), .number(1000)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .hsla, arguments: [.number(0), .number(0), .number(0), .number(-1)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .hsla, arguments: [.number(0), .number(0), .number(0), .number(1000)])))
     }
 
     func testExpressionInitFailureTooFewArguments() {
-        XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: Array(repeating: .number(0), count: .random(in: 0...3)))))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .rgb, arguments: [.number(0), .number(0)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: [.number(0), .number(0), .number(9)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .hsl, arguments: [.number(0), .number(0)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .hsla, arguments: [.number(0), .number(0), .number(9)])))
     }
 
     func testExpressionInitFailureTooManyArguments() {
-        XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: Array(repeating: .number(0), count: .random(in: 5...20)))))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .rgb, arguments: [.number(0), .number(0), .number(9), .number(9)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .rgba, arguments: [.number(0), .number(0), .number(9), .number(3), .number(3)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .hsl, arguments: [.number(0), .number(0), .number(9), .number(9)])))
+        XCTAssertNil(StyleColor(expression: Exp(operator: .hsla, arguments: [.number(0), .number(0), .number(9), .number(3), .number(3)])))
     }
 
-    func testRGBAStringInit() {
-        func randomSpaces() -> String {
-            String(repeating: " ", count: .random(in: 0...100))
-        }
-        verify(StyleColor(rgbaString: "\(randomSpaces())rgba(\(randomSpaces())\(red!)\(randomSpaces()),\(randomSpaces())\(green!)\(randomSpaces()),\(randomSpaces())\(blue!)\(randomSpaces()),\(randomSpaces())\(alpha!)\(randomSpaces()))\(randomSpaces())"))
+    func testStringInit() {
+        let colorString = "rgb(1,2,3)"
+        XCTAssertEqual(StyleColor(rawValue: colorString).rawValue, colorString)
     }
 
-    func testRGBAStringInitNoDecimals() {
-        red = 1
-        green = 2
-        blue = 3
-        alpha = 0
-        verify(StyleColor(rgbaString: "rgba(1,2,3,0)"))
-    }
-
-    func testRGBAStringInitFailureInvalidStructure() {
-        XCTAssertNil(StyleColor(rgbaString: ""))
-        XCTAssertNil(StyleColor(rgbaString: "rgba (0, 0, 0, 0)"))
-        XCTAssertNil(StyleColor(rgbaString: "rgb(0, 0, 0)"))
-        XCTAssertNil(StyleColor(rgbaString: "rgb(0, 0, 0, 0)"))
-        XCTAssertNil(StyleColor(rgbaString: "rgba(0, 0, 0)"))
-        XCTAssertNil(StyleColor(rgbaString: "rgba(a, b, c)"))
-        XCTAssertNil(StyleColor(rgbaString: "rgba(0.0.0, 0.0.0, 0.0.0, 0.0.0)"))
-        XCTAssertNil(StyleColor(rgbaString: "abcdrgba(1,2,3,0)"))
-        XCTAssertNil(StyleColor(rgbaString: "rgba(1,2,3,0)abcd"))
-    }
-
-    func testRGBAStringInitFailureValuesOutOfBounds() {
-        func makeRGBAString(red: Double = 0, green: Double = 0, blue: Double = 0, alpha: Double = 0) -> String {
-            "rgba(\(red),\(green),\(blue),\(alpha))"
-        }
-        XCTAssertNil(StyleColor(rgbaString: makeRGBAString(red: invalidComponentValue(outsideOf: 0...255))))
-        XCTAssertNil(StyleColor(rgbaString: makeRGBAString(green: invalidComponentValue(outsideOf: 0...255))))
-        XCTAssertNil(StyleColor(rgbaString: makeRGBAString(blue: invalidComponentValue(outsideOf: 0...255))))
-        XCTAssertNil(StyleColor(rgbaString: makeRGBAString(alpha: invalidComponentValue(outsideOf: 0...1))))
-    }
-
-    func testRGBAString() {
-        let styleColor = StyleColor(red: red, green: green, blue: blue, alpha: alpha)
-
-        XCTAssertEqual(styleColor?.rgbaString, rgbaString)
+    func testLiteralInit() {
+        XCTAssertEqual("rgb(1,2,3)" as StyleColor, "rgb(1,2,3)")
+        XCTAssertEqual("rgb(\(1),\(2),\(3))" as StyleColor, "rgb(1,2,3)")
     }
 
     func testEquatable() {
         XCTAssertEqual(
+            StyleColor(red: red, green: green, blue: blue),
+            StyleColor(red: red, green: green, blue: blue)
+        )
+        XCTAssertEqual(
             StyleColor(red: red, green: green, blue: blue, alpha: alpha),
-            StyleColor(red: red, green: green, blue: blue, alpha: alpha))
+            StyleColor(red: red, green: green, blue: blue, alpha: alpha)
+        )
+        XCTAssertEqual(
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness),
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness)
+        )
+        XCTAssertEqual(
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness, alpha: alpha),
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness, alpha: alpha)
+        )
     }
 
     func testCodableRoundtrip() throws {
-        let color = StyleColor(red: red, green: green, blue: blue, alpha: alpha)!
+        let colors = [
+            StyleColor(red: red, green: green, blue: blue)!,
+            StyleColor(red: red, green: green, blue: blue, alpha: alpha)!,
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness)!,
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness, alpha: alpha)!
+        ]
+        for color in colors {
+            // wrapping in an array since iOS 12 and lower only support
+            // Array and Dictionary as the top level JSON values
+            let data = try JSONEncoder().encode([color])
 
-        // wrapping in an array since iOS 12 and lower only support
-        // Array and Dictionary as the top level JSON values
-        let data = try JSONEncoder().encode([color])
+            let decodedColor = try JSONDecoder().decode([StyleColor].self, from: data)
 
-        let decodedColor = try JSONDecoder().decode([StyleColor].self, from: data)
-
-        XCTAssertEqual(decodedColor, [color])
+            XCTAssertEqual(decodedColor, [color])
+        }
     }
 
     func testEncoding() throws {
-        let color = StyleColor(red: red, green: green, blue: blue, alpha: alpha)!
+        let colors = [
+            StyleColor(red: red, green: green, blue: blue)!,
+            StyleColor(red: red, green: green, blue: blue, alpha: alpha)!,
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness)!,
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness, alpha: alpha)!
+        ]
+        let expectedColorStrings = [rgbString, rgbaString, hslString, hslaString]
 
-        // wrapping in an array since iOS 12 and lower only support
-        // Array and Dictionary as the top level JSON values
-        let data = try JSONEncoder().encode([color])
+        for (color, expectedColorString) in zip(colors, expectedColorStrings) {
+            // wrapping in an array since iOS 12 and lower only support
+            // Array and Dictionary as the top level JSON values
+            let data = try JSONEncoder().encode([color])
 
-        // Double quotes are added to the expected value to make it a JSON string
-        XCTAssertEqual(data, #"["\#(rgbaString)"]"#.data(using: .utf8))
+            // Double quotes are added to the expected value to make it a JSON string
+            XCTAssertEqual(data, #"["\#(expectedColorString)"]"#.data(using: .utf8))
+        }
     }
 
-    func testDecodingRGBAExpression() throws {
-        let expressionData = #"["rgba",\#(red!),\#(green!),\#(blue!),\#(alpha!)]"#.data(using: .utf8)!
+    func testDecodingExpression() throws {
+        let rawExpressions = [
+            #"["rgb",\#(red!),\#(green!),\#(blue!)]"#,
+            #"["rgba",\#(red!),\#(green!),\#(blue!),\#(alpha!)]"#,
+            #"["hsl",\#(hue!),\#(saturation!),\#(lightness!)]"#,
+            #"["hsla",\#(hue!),\#(saturation!),\#(lightness!),\#(alpha!)]"#
+        ]
+        let expectedColorStrings = [rgbString, rgbaString, hslString, hslaString]
 
-        let color = try JSONDecoder().decode(StyleColor.self, from: expressionData)
+        for (rawExpression, colorString) in zip(rawExpressions, expectedColorStrings) {
+            let color = try JSONDecoder().decode(StyleColor.self, from: rawExpression.data(using: .utf8)!)
 
-        // Accuracy can be removed if https://bugs.swift.org/browse/SR-15172 is fixed
-        XCTAssertEqual(color.red, red!, accuracy: 0.001)
-        XCTAssertEqual(color.green, green!, accuracy: 0.001)
-        XCTAssertEqual(color.blue, blue!, accuracy: 0.001)
-        XCTAssertEqual(color.alpha, alpha!, accuracy: 0.001)
+            XCTAssertEqual(color.rawValue, colorString)
+        }
     }
 
-    func testDecodingRGBAString() throws {
-        // wrapping in an array since iOS 12 and lower only support
-        // Array and Dictionary as the top level JSON values
-        // Double quotes are added to make it a JSON string
-        let rgbaJSONString = #"["\#(rgbaString)"]"#.data(using: .utf8)!
+    func testDecodingString() throws {
+        let colorStrings = [rgbString, rgbaString, hslString, hslaString]
+        let expectedColors = [
+            StyleColor(red: red, green: green, blue: blue)!,
+            StyleColor(red: red, green: green, blue: blue, alpha: alpha)!,
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness)!,
+            StyleColor(hue: hue, saturation: saturation, lightness: lightness, alpha: alpha)!
+        ]
 
-        let color = try JSONDecoder().decode([StyleColor].self, from: rgbaJSONString)
+        for (colorString, expectedColor) in zip(colorStrings, expectedColors) {
+            // wrapping in an array since iOS 12 and lower only support
+            // Array and Dictionary as the top level JSON values
+            // Double quotes are added to make it a JSON string
+            let rgbaJSONString = #"["\#(colorString)"]"#.data(using: .utf8)!
 
-        XCTAssertEqual(color, [StyleColor(red: red, green: green, blue: blue, alpha: alpha)!])
+            let color = try JSONDecoder().decode([StyleColor].self, from: rgbaJSONString)
+
+            XCTAssertEqual(color, [expectedColor])
+        }
     }
 
-    func testColorSpacesWithStandardRangeColorValues() throws {
+        func testColorSpacesWithStandardRangeColorValues() throws {
         // given
         let uiColors = colorSpaces
             .map { CGColor(colorSpace: $0, components: [0.5, 0.5, 0.5, 0.5, 0.5])! }
@@ -267,9 +305,13 @@ final class StyleColorTests: XCTestCase {
         let styleColor = StyleColor(faultyUIColor)
 
         // then
-        XCTAssertEqual(styleColor.red, 0)
-        XCTAssertEqual(styleColor.green, 0)
-        XCTAssertEqual(styleColor.blue, 0)
-        XCTAssertEqual(styleColor.alpha, 1)
+        XCTAssertEqual(styleColor.rawValue, "rgba(0.00, 0.00, 0.00, 1.00)")
+    }
+}
+
+extension Array {
+    func rotatingLeft(positions: Int) -> Self {
+        let index = index(startIndex, offsetBy: positions, limitedBy: endIndex) ?? endIndex
+        return Array(self[index...] + self[..<index])
     }
 }