+++ CHANGELOG.md
@@ -4,7 +4,14 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
+### Breaking changes ⚠️
+
+* `StyleColor.red`, `StyleColor.green`, `StyleColor.blue`, `StyleColor.alpha` are not in use anymore and got removed.
+
+### Features ✨ and improvements 🏁
+
 * Add MSAA support with the `MapInitOptions/antialiasingSampleCount` property.
+* `StyleColor` - add support for all color formats as defined by [Mapbox Style Spec](https://docs.mapbox.com/style-spec/reference/types/#color).
 
 ## 11.0.0-beta.3 - 8 September, 2023